requirements.txt
indic_transliteration
curation_utils
doc_curation
vidyut-prakriya
vidyut-lipi

